refactor acs-refresh-dev.py to reduce special local setup
