+++ util.py
@@ -1,4 +1,6 @@
 import os, errno
+import yaml
+
 def make_dirs(path):
     try:
         os.makedirs(path)
@@ -7,3 +9,21 @@ def make_dirs(path):
             pass
         else:
             raise
+
+# load conf file (yml)
+def getConfig(filename, stage='dev'):
+    # sanity check
+    if not filename:
+        return None
+
+    ret = None
+    with open(filename, 'r') as file:
+        conf = yaml.load(file)
+        ret = conf
+        if conf['default']:
+            ret = conf['default']
+            if stage and stage in conf:  # override default
+                for key in conf[stage]:
+                    ret[key] = conf[stage][key]
+
+    return ret;
